solucion al problema del SSR(Recomendable no usarlo)

+++ src/app/services/token.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 const TOKEN_KEY = 'AuthToken';
 const USERNAME_KEY = 'AuthUserName';
 const AUTHORITIES_KEY = 'AuthAuthorities';
-const ID_ENTIDAD_KEY = 'AuthIdEntidad';  // Nueva constante para idEntidad
+const ID_ENTIDAD_KEY = 'AuthIdEntidad';
 
 @Injectable({
   providedIn: 'root'
@@ -14,32 +14,42 @@ export class TokenService {
 
   constructor() { }
 
+  private isBrowser(): boolean {
+    return typeof window !== 'undefined';
+  }
+
   public setToken(token: string): void {
-    window.sessionStorage.removeItem(TOKEN_KEY);
-    window.sessionStorage.setItem(TOKEN_KEY, token);
+    if (this.isBrowser()) {
+      window.sessionStorage.removeItem(TOKEN_KEY);
+      window.sessionStorage.setItem(TOKEN_KEY, token);
+    }
   }
 
   public getToken(): string | null {
-    return sessionStorage.getItem(TOKEN_KEY);
+    return this.isBrowser() ? window.sessionStorage.getItem(TOKEN_KEY) : null;
   }
 
   public setUserName(userName: string): void {
-    window.sessionStorage.removeItem(USERNAME_KEY);
-    window.sessionStorage.setItem(USERNAME_KEY, userName);
+    if (this.isBrowser()) {
+      window.sessionStorage.removeItem(USERNAME_KEY);
+      window.sessionStorage.setItem(USERNAME_KEY, userName);
+    }
   }
 
   public getUserName(): string | null {
-    return sessionStorage.getItem(USERNAME_KEY);
+    return this.isBrowser() ? window.sessionStorage.getItem(USERNAME_KEY) : null;
   }
 
   public setAuthorities(authorities: string[]): void {
-    window.sessionStorage.removeItem(AUTHORITIES_KEY);
-    window.sessionStorage.setItem(AUTHORITIES_KEY, JSON.stringify(authorities));
+    if (this.isBrowser()) {
+      window.sessionStorage.removeItem(AUTHORITIES_KEY);
+      window.sessionStorage.setItem(AUTHORITIES_KEY, JSON.stringify(authorities));
+    }
   }
 
   public getAuthorities(): string[] {
     this.roles = [];
-    if (sessionStorage.getItem(AUTHORITIES_KEY)) {
+    if (this.isBrowser() && sessionStorage.getItem(AUTHORITIES_KEY)) {
       JSON.parse(sessionStorage.getItem(AUTHORITIES_KEY)!).forEach((authority: any) => {
         this.roles.push(authority.authority);
       });
@@ -47,21 +57,23 @@ export class TokenService {
     return this.roles;
   }
 
- 
-
   public setIdEntidad(idEntidad: number): void {
-    if (idEntidad != null) {
+    if (this.isBrowser() && idEntidad != null) {
       sessionStorage.setItem(ID_ENTIDAD_KEY, idEntidad.toString());
     }
   }
 
-  // Obtener el ID de la entidad
   public getIdEntidad(): number | null {
-    const idEntidad = sessionStorage.getItem(ID_ENTIDAD_KEY);
-    return idEntidad ? parseInt(idEntidad) : null;
+    if (this.isBrowser()) {
+      const idEntidad = sessionStorage.getItem(ID_ENTIDAD_KEY);
+      return idEntidad ? parseInt(idEntidad) : null;
+    }
+    return null;
   }
 
   public logOut(): void {
-    window.sessionStorage.clear();
+    if (this.isBrowser()) {
+      window.sessionStorage.clear();
+    }
   }
 }
